+++ src/App.svelte
@@ -1,20 +1,20 @@
 <script lang="ts">
+    import { loadMovieData, loadHistoryData, storedActiveSelection } from './stores';
+    import { ID_MANHATTAN, ID_NETHERLAND, ID_MOVIE, ID_BURG_BLAUSTEIN, UNLOADED, LOADED } from "./_interfaces";
     import { ApolloClient, InMemoryCache } from "@apollo/client";
     import { setClient } from "svelte-apollo";
     import Welcome from "./comps/Welcome.svelte";
-    import Hearts from "./comps/Hearts.svelte";
-    import Changes from "./comps/Changes.svelte";
+    import Hearts from "./comps/Specials/Hearts.svelte";
+    import Changes from "./comps/Product/Changes.svelte";
     import Filter from "./comps/Filter/Filter.svelte";
     import Manhattan from "./comps/Specials/Manhattan.svelte";
     import Netherland from "./comps/Specials/Netherland.svelte";
     import Blaustein from "./comps/Specials/Blaustein.svelte";
     import Products from "./comps/Product/Products.svelte";
-    import { loadMovieData, loadHistoryData, storedActiveSelection } from './stores';
-    import { ID_MANHATTAN, ID_NETHERLAND, ID_MOVIE, ID_BURG_BLAUSTEIN, UNLOADED, AFF_LINK, LOADED } from "./_interfaces";
-    import Icon from "./comps/Icon.svelte";
-    import Notifications from "./comps/Notifications.svelte";
-    import Github from "./comps/Github.svelte";
-    import Darkmode from "./comps/Darkmode.svelte";
+    import Notifications from "./comps/Features/Notifications.svelte";
+    import Github from "./comps/Features/Github.svelte";
+    import Darkmode from "./comps/Features/Darkmode.svelte";
+    import Footer from "./comps/Footer.svelte";
 
     const client = new ApolloClient({
         uri: 'https://api.bbx.watch/api/graphql',
@@ -24,14 +24,12 @@
     setClient(client);
 
     let activeTagIds;
-    let lastCursor;
     let loadedData;
 
     loadHistoryData()
 
     storedActiveSelection.subscribe(store => {
         activeTagIds = store.tags;
-        lastCursor = store.lastCursor;
         loadedData = store.loadedData;
 
         // load movie data
@@ -65,33 +63,22 @@
             <Blaustein/>
         {/if}
         <Products/>
-
     {/if}
 
-    <div class="footer">
-        <strong>Legende:</strong><br/>
-        <Icon modifier="new"/>
-        = Neues Produkt
-        <Icon modifier="flame"/>
-        = Beliebtes Produkt (mehr als 2 mal Verfügbar)
-        <Icon modifier="heart" svg="true" class="active"/>
-        = "Will Ich haben" Produkt<br/>
-        <span class="state state--blue"></span> = Verfügbar
-        <span class="state state--green"></span> = Bald erhältlich
-        <span class="state state--red"></span> = Zurzeit Vergriffen
-        <span class="state state--orange"></span> = In Ankündigung
-        <br/><br/>
-        <strong>Stand:</strong> {lastCursor.dateStr}
-        {#if !!AFF_LINK}
-        <div class="small">
-            * Die ausgehenden Produktlinks sind Affiliate-Links. Wenn Ihr klickt und kauft, dann unterstütz Ihr dieses Projekt. Vielen Dank :)
-        </div>
-        {/if}
-    </div>
+    <Footer />
 </main>
 
 <style lang="scss">
   @import './scss/variables';
+
+  main {
+    @media (min-width: 1024px) {
+      width: 65vw;
+    }
+
+    padding: 0 $space-lg 300px $space-lg;
+    margin: 0 auto;
+  }
 
   .notice {
     font-weight: bold;
@@ -113,6 +100,9 @@
     font-family: $font-family;
     font-size: $base-font-size;
     overflow-y: scroll;
+    margin: 0;
+    // chrome blue touch highlight
+    -webkit-tap-highlight-color: transparent;
   }
 
   // test darkmode https://stackoverflow.com/questions/57606960/how-can-i-emulate-prefers-color-scheme-media-query-in-chrome
@@ -197,10 +187,6 @@
     text-align: right;
   }
 
-  :global(.filter-headline) {
-    font-size: ms(1);
-  }
-
   :global(.flex.bl) {
     position: relative;
   }
@@ -214,38 +200,13 @@
     border-left: solid 1px $color-primary;
   }
 
-  main {
-    @media (min-width: 1024px) {
-      width: 65vw;
-    }
+  :global(h4.tag-name) {
+    transform: rotate(-90deg);
+    width: 20px;
 
-    padding-bottom: 300px;
-    margin: 0 auto;
-
-    .footer {
-      padding-top: 50px;
-      font-size: ms(-1);
-    }
-
-    .state {
-      padding: $space-lg;
-      border-radius: 100%;
-      background: $color-primary;
-      font-size: 0;
-      vertical-align: middle;
-      margin-right: $space-xs;
-
-      &--green {
-        background: $color-comingsoon;
-      }
-
-      &--red {
-        background: $color-unavailable;
-      }
-
-      &--orange {
-        background: $color-annoucement;
-      }
+    @media (min-width: 400px) {
+      transform: rotate(0);
+      width: 60px;
     }
   }
 </style>