+++ src/comps/Filter/FilterStates.svelte
@@ -112,7 +112,7 @@
 </script>
 
 <div class="flex">
-    <h4>Status</h4>
+    <h4 class="tag-name">Status</h4>
     <div class="flex flex--wrap bl">
         {#each sortedItems as state (state.id)}
             <div class={getClasses(state)}